requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
psycopg2
mysqlclient
django-environ
